Avoid unnecessary logic once current_user iset set.

+++ app/controllers/application_controller.rb
@@ -23,6 +23,7 @@ class ApplicationController < ActionController::Base
   end
 
   def current_user
+    return @current_user if defined? @current_user
     @current_user = current_user_session && current_user_session.record
     if @current_user && ( @current_user.net_id != sso_net_id )
       current_user_session.destroy
